server/pyproject.toml
[project]
name = "drf-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "coverage>=7.9.2",
    "django>=5.2.4",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.7.0",
    "django-debug-toolbar>=6.0.0",
    "django-environ>=0.12.0",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.94.2",
    "drf-spectacular>=0.28.0",
    "isort>=6.0.1",
    "mypy>=1.17.0",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.9",
    "pytest-django>=4.11.1",
    "ruff>=0.12.3",
]
